Use right size for user.image and extract submitted by logic
The user.image is used directly on all cases and just re-sizeing it with css. The default size of the image is 460x460 and the project uses 32x32 and 16x16 so in favor of use the right size for browser performance, the `user_image_url` helper was created  to attach the  `s` query parameter to the URL so we can request the image in the size we want.

The second change is the extraction of the submitted by logic. That logic was spread in three different files, so the best option to avoid copy/paste is to move it to a helper (`gemfile_submitted_by`) and re-use it when it's needed it.